Add post call to register
