+++ src/store.js
@@ -1,6 +1,6 @@
 import { createStore, applyMiddleware, combineReducers } from 'redux';
 import thunk from 'redux-thunk';
-import registerStore from './register/registerStore.tsx';
+import registerStore from './register/registerStore.ts';
 
 const store = createStore(
   combineReducers({
